src/requirements.txt
osmnx==0.10
PyYAML==5.1
Rtree==0.8.3
